requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pymongo==4.5.0
python-dotenv==1.0.0
apscheduler==3.10.1
dnspython==2.4.2
loguru==0.7.0
flask==2.3.3
waitress==2.1.2